pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm[toml]"]
build-backend = "setuptools.build_meta"

[project]
dependencies = [
    'girder-client',
    'numpy',
    'setuptools',
    'opencv-python',
    'pandas',
    'pymongo',
    'pillow',
    'dash',
    'dash_bootstrap_components'
]
name = "dsa_helpers"
authors = [
    {name = "Juan C. Vizcarra", email = "jvizcar@emory.edu"},
]
maintainers = [
    {name = "Juan C. Vizcarra", email = "jvizcar@emory.edu"}
]
description = "Utility functions for working with the DSA girder client."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
keywords = ["dsa", "girder", "girder_client"]
version = "v0.3dev1"

[project.urls]
Repository = "https://github.com/Gutman-Lab/dsa-python-utils"
